complete department management with error handling

+++ src/gui/DepartmentListController.java
@@ -161,19 +161,20 @@ public class DepartmentListController implements Initializable, DataChangeListen
 		}
 
 	private void removeEntity(Department obj) {
-		Optional<ButtonType> result = Alerts.showConfirmation("Confirmation", "Are you sure to delete?");
-		
-		if (result.get() == ButtonType.OK) {
-			if (service == null) {
-				throw new IllegalStateException("Service was null");
-			}
-			try {
-				service.remove(obj);
-				updateTableView();
-			}
-			catch(DbIntegrityException e) {
-				Alerts.showAlert("Error removing object", null, e.getMessage(), AlertType.ERROR);
-				}
-			}
-	} 
+	    Optional<ButtonType> result = Alerts.showConfirmation("Confirmation", "Are you sure to delete?");
+	    
+	    if (result.isPresent() && result.get() == ButtonType.OK) {
+	        if (service == null) {
+	            throw new IllegalStateException("Service was null");
+	        }
+	        try {
+	            service.remove(obj);
+	            updateTableView();
+	        }
+	        catch(DbIntegrityException e) {
+	            Alerts.showAlert("Error removing object", "Integrity error", 
+	                "You cannot delete a department that has sellers linked to it.", AlertType.ERROR);
+	        }
+	    }
+	}
 }
